requirements.txt
asgiref==3.7.2
astroid==2.15.6
black==23.7.0
click==8.1.4
dill==0.3.6
Django==4.2.3
djangorestframework==3.14.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.1
pylint==2.17.4
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.7.1
wrapt==1.15.0
